# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Takas Backend API",
    "description": "İkinci el takas mobil uygulaması için REST API",
    "version": "1.0.0",
    "contact": {
      "name": "Takas Backend Team",
      "email": "info@takas.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "username": {"type": "string"},
          "email": {"type": "string"},
          "first_name": {"type": "string"},
          "last_name": {"type": "string"},
          "phone": {"type": "string"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Kullanıcı kaydı",
        "description": "Yeni kullanıcı hesabı oluşturur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {"type": "string", "example": "johndoe"},
                  "email": {"type": "string", "format": "email", "example": "john@example.com"},
                  "password": {"type": "string", "example": "password123"},
                  "first_name": {"type": "string", "example": "John"},
                  "last_name": {"type": "string", "example": "Doe"},
                  "phone": {"type": "string", "example": "+905551234567"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kullanıcı başarıyla oluşturuldu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "400": {"description": "Geçersiz veri"},
          "500": {"description": "Sunucu hatası"}
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Kullanıcı girişi",
        "description": "Kullanıcı kimlik doğrulaması yapar ve JWT token döner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {"type": "string", "example": "johndoe"},
                  "password": {"type": "string", "example": "password123"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Giriş başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "access_token": {"type": "string"},
                    "refresh_token": {"type": "string"},
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {"description": "Geçersiz kimlik bilgileri"},
          "500": {"description": "Sunucu hatası"}
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Kullanıcı profili",
        "description": "Giriş yapmış kullanıcının profil bilgilerini döner",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Profil bilgileri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {"description": "Yetkilendirme gerekli"},
          "404": {"description": "Kullanıcı bulunamadı"}
        }
      }
    },
...
  },
...
}